Switch all terrain shaders to unroll the per-world loop based on dispatch dimension .y.
This appears to be significantly faster than having a "persistent" kernel with a for-loop over this dimension. The reason for the speedup is probably that gl_WorkGroupID.y is block-level uniform, so the compiler has some extra information. I guess the dispatch ordering will also be better - an entire world will be dispatched at once, causing that region of memory to be focused on.