Solve slow shader compilation due to Beats.glsl
The problem was the macro expansion was causing a huge volume of text to come out of the preprocessing stage.

I did not realize this right away. I actually wasted several days thinking about this problem, I thought that parsing the shader was fundamentally too slow. A couple of solutions I came up with were:
1. Multi threaded shader compilation. Would not help that much because the main issue was a single shader taking 5 seconds to load, and the others all being pretty fast (100ms or less).
2. Specialization regions, compiling shaders where beats.glsl is mostly #ifdef'd off, so the shader is only valid for a sub-section of the world. This would have definition solve the problem, but adds a ton of complexity selecting the correct shader and dealing with that.
3. SPIR-V specialization. Using SPIR-V tools to separately build and link the beats function from the rest of the shader. This only solves the problem when beats.glsl doesn't change though. And it's unclear if it would ever solve the problem because building final shader is likely where the overall expense remains.

In the end I tested using regular functions instead of macros and that completely solved the slow compilation (5 seconds down to 0.5s). I guess the real issue wasn't compilation, it was parsing the giant volume of text coming out of the macro functions.